D49: the numeric kinds named as XSLT 2.0 names them, integer and double
TypedValue.Int is Integer and TypedValue.Real is Double; the kinds do not
change. The cast vocabulary gains integer and double beside number, on
operands, sorts, min and max now and on capture bindings when design 25
section 9 lands (D50). Stroom's Val widths, and the plain words integer and
decimal, were weighed and the decision says why: decimal is XSLT's exact kind
and stays free for one. Designs 16, 17 and 25 name the kinds as the code
does. Engine 566, pipeline 154, app 5, checkstyle clean; no golden moved.

+++ stroom-shapeshifter/stroom-shapeshifter-engine/src/test/java/stroom/shapeshifter/engine/value/DatesTest.java
@@ -68,7 +68,7 @@ class DatesTest {
         assertThat(Comparisons.compare(paris, utc)).isZero();
         // Date arithmetic through the millis cast: a duration is a subtraction away.
         assertThat(Transforms.subtract(List.of(paris, utc)))
-                .isEqualTo(new TypedValue.Int(0));
+                .isEqualTo(new TypedValue.Integer(0));
     }
 
     @Test
